AuthService/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose
pydantic
pydantic[email]
sqlalchemy
alembic
pytest
httpx
pytest-asyncio
psycopg2-binary
